MAS API Provider
