+++ shared/domain/src/main/java/gov/va/vro/model/mas/MasCollectionStatus.java
@@ -1,4 +1,4 @@
-package gov.va.vro.service.provider.mas.model;
+package gov.va.vro.model.mas;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
