+++ shared/domain/src/main/java/gov/va/vro/model/mas/MasCollectionStatusReq.java
@@ -1,4 +1,4 @@
-package gov.va.vro.service.provider.mas.model;
+package gov.va.vro.model.mas;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -6,12 +6,11 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-/** @author warren @Date 10/11/22 */
-@NoArgsConstructor
 @Getter
 @Setter
+@NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class MasOrderExam {
-  @JsonProperty("response")
-  private String response;
+public class MasCollectionStatusReq {
+  @JsonProperty("collectionsId")
+  private int collectionsId;
 }
