+++ shared/domain/src/main/java/gov/va/vro/model/mas/MasDocument.java
@@ -1,0 +1,25 @@
+package gov.va.vro.model.mas;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+/** @author warren @Date 10/5/22 */
+@NoArgsConstructor
+@Getter
+@Setter
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class MasDocument {
+  @JsonProperty("eFolderVersionRefId")
+  private String efolderversionrefid;
+
+  @JsonProperty("condition")
+  private String condition;
+
+  @JsonProperty("annotations")
+  private List<MasAnnotation> annotations;
+}
