+++ shared/domain/src/main/java/gov/va/vro/model/mas/MasOrderExamConditions.java
@@ -1,4 +1,4 @@
-package gov.va.vro.service.provider.mas.model;
+package gov.va.vro.model.mas;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -6,17 +6,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.List;
-
 /** @author warren @Date 10/11/22 */
 @NoArgsConstructor
 @Getter
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class MasOrderExamReq {
-  @JsonProperty("conditions")
-  private List<MasOrderExamConditions> conditions;
+public class MasOrderExamConditions {
+  @JsonProperty("contentionText")
+  private String contentionText;
 
-  @JsonProperty("collectionsId")
-  private Integer collectionsId;
+  @JsonProperty("conditionCode")
+  private String conditionCode;
 }
