+++ shared/domain/src/main/java/gov/va/vro/model/mas/MasStatus.java
@@ -1,4 +1,4 @@
-package gov.va.vro.service.provider.mas.model;
+package gov.va.vro.model.mas;
 
 import java.util.HashMap;
 import java.util.Map;
